Add jwt auth with devise
The authentication of request using user email and api key is done using jwt.
Two warden strategies were implemented and added to devise:
- one for getting an access token at users sign_in API endpoint (token_gen_strategy.rb)
- the other for authentication with the received token (jwt_strategy.rb)

So, you have to make a POST request with user credentials in the request body, in json format:
{email: 'email', password: 'pass'}
On successful sign in, you get a token, by which protected enpoints can be accessed.